PP-6828 Organise cypress stubs
 - Moves stub helpers to `stubs` folder
