+++ test/cypress/integration/dashboard/dashboard-links.cy.test.js
@@ -1,8 +1,8 @@
 'use strict'
 
-const userStubs = require('../../utils/user-stubs')
-const gatewayAccountStubs = require('../../utils/gateway-account-stubs')
-const transactionsSummaryStubs = require('../../utils/transaction-summary-stubs')
+const userStubs = require('../../stubs/user-stubs')
+const gatewayAccountStubs = require('../../stubs/gateway-account-stubs')
+const transactionsSummaryStubs = require('../../stubs/transaction-summary-stubs')
 
 describe('the links are displayed correctly on the dashboard', () => {
   const userExternalId = 'cd0fa54cf3b7408a80ae2f1b93e7c16e'
