+++ test/cypress/integration/manage-team-members/edit-permissions.cy.test.js
@@ -1,7 +1,7 @@
 'use strict'
 
-const userStubs = require('../../utils/user-stubs')
-const inviteStubs = require('../../utils/invite-stubs')
+const userStubs = require('../../stubs/user-stubs')
+const inviteStubs = require('../../stubs/invite-stubs')
 
 const SERVICE_EXTERNAL_ID = 'service_abc_123'
 const AUTHENTICATED_USER_ID = 'authenticated-user-id'
