+++ test/cypress/integration/partnerapp/handle-redirect-to-gocardless.cy.test.js
@@ -1,6 +1,6 @@
-const gatewayAccountStubs = require('../../utils/gateway-account-stubs')
-const userStubs = require('../../utils/user-stubs')
-const goCardlessStubs = require('../../utils/go-cardless-connect-stubs')
+const gatewayAccountStubs = require('../../stubs/gateway-account-stubs')
+const userStubs = require('../../stubs/user-stubs')
+const goCardlessStubs = require('../../stubs/go-cardless-connect-stubs')
 
 describe('Connect to Go Cardless', () => {
   const userExternalId = 'cd0fa54cf3b7408a80ae2f1b93e7c16e'
