+++ test/cypress/integration/payment-links/create-payment-link.cy.test.js
@@ -1,5 +1,5 @@
-const userStubs = require('../../utils/user-stubs')
-const gatewayAccountStubs = require('../../utils/gateway-account-stubs')
+const userStubs = require('../../stubs/user-stubs')
+const gatewayAccountStubs = require('../../stubs/gateway-account-stubs')
 const userExternalId = 'a-user-id'
 const gatewayAccountId = 42
 const serviceName = {
