+++ test/cypress/integration/payment-links/edit-payment-link.cy.test.js
@@ -1,6 +1,6 @@
-const userStubs = require('../../utils/user-stubs')
-const gatewayAccountStubs = require('../../utils/gateway-account-stubs')
-const { getProductsStub, getProductByExternalIdStub } = require('../../utils/products-stubs')
+const userStubs = require('../../stubs/user-stubs')
+const gatewayAccountStubs = require('../../stubs/gateway-account-stubs')
+const { getProductsStub, getProductByExternalIdStub } = require('../../stubs/products-stubs')
 const userExternalId = 'a-user-id'
 const gatewayAccountId = 42
