+++ test/cypress/integration/request-to-go-live/agreement.cy.test.js
@@ -3,9 +3,9 @@
 const lodash = require('lodash')
 const utils = require('../../utils/request-to-go-live-utils')
 const { userExternalId, gatewayAccountId, serviceExternalId } = utils.variables
-const gatewayAccountStubs = require('../../utils/gateway-account-stubs')
-const userStubs = require('../../utils/user-stubs')
-const goLiveRequestStubs = require('../../utils/go-live-request-stubs')
+const gatewayAccountStubs = require('../../stubs/gateway-account-stubs')
+const userStubs = require('../../stubs/user-stubs')
+const goLiveRequestStubs = require('../../stubs/go-live-request-stubs')
 
 describe('Request to go live: agreement', () => {
   beforeEach(() => {
