+++ test/cypress/integration/request-to-go-live/organisation-name.cy.test.js
@@ -1,7 +1,7 @@
 'use strict'
 
 const utils = require('../../utils/request-to-go-live-utils')
-const serviceStubs = require('../../utils/service-stubs')
+const serviceStubs = require('../../stubs/service-stubs')
 const { userExternalId, gatewayAccountId, serviceExternalId } = utils.variables
 const notStartedServiceRole = utils.buildServiceRoleForGoLiveStage('NOT_STARTED')
 const requestToGoLivePageOrganisationNameUrl = `/service/${serviceExternalId}/request-to-go-live/organisation-name`
