+++ test/cypress/integration/transactions/transaction-search.cy.test.js
@@ -1,7 +1,7 @@
-const userStubs = require('../../utils/user-stubs')
-const gatewayAccountStubs = require('../../utils/gateway-account-stubs')
-const stripeAccountSetupStubs = require('../../utils/stripe-account-setup-stub')
-const transactionsStubs = require('../../utils/transaction-stubs')
+const userStubs = require('../../stubs/user-stubs')
+const gatewayAccountStubs = require('../../stubs/gateway-account-stubs')
+const stripeAccountSetupStubs = require('../../stubs/stripe-account-setup-stub')
+const transactionsStubs = require('../../stubs/transaction-stubs')
 
 const transactionsUrl = `/transactions`
 const userExternalId = 'cd0fa54cf3b7408a80ae2f1b93e7c16e'
