+++ test/cypress/stubs/invite-stubs.js
