+++ test/cypress/stubs/products-stubs.js
