+++ test/cypress/stubs/stripe-account-setup-stub.js
