+++ test/cypress/stubs/transaction-summary-stubs.js
